endlessWorld/collage_nav/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
branca==0.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
crispy-bootstrap5==2025.4
Django==5.2.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
folium==0.19.6
frozenlist==1.6.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
numpy==2.2.6
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
twilio==9.6.1
urllib3==2.4.0
xyzservices==2025.4.0
yarl==1.20.0
